Move J---- to `java-bindings`
