+++ android/src/main/cpp/java-bindings/JArrayList.h
