+++ android/src/main/cpp/java-bindings/JHashMap.cpp
