+++ android/src/main/cpp/java-bindings/JHashMap.h
