+++ android/src/main/cpp/java-bindings/JImageProxy.h
