+++ android/src/main/cpp/java-bindings/JImageProxyHostObject.cpp
