+++ android/src/main/cpp/java-bindings/JImageProxyHostObject.h
