+++ android/src/main/cpp/java-bindings/JReadableArray.h
